Added tests for shutterWindow()
They are a bit ugly for the moment, I shall probably compress these down
into a single test in the future, probably table driven i would venture.

I see the fact that these tests take up over 50 percent of the test list
to be an idication of multiple concerns whithin the class.