Fixed new segment no repair to start bug
The bug was that when you started a new segment
you couldn't repair the window you started on
until you moved away from it and back again.

Upon inspection i discovered problem was that the
referance to the new segment was being grabbed
after we moved to the start and grabbed a referance
to the start window, meaning the window we were
grabbing was on the old segment.

To fix i moved the call to aquireSegmentReferance to
the start of resetFelix, before the call to
moveToWindow.