Started TDD on MultiDimentionalArray
Current have tests to ensure that it can remember
it's row and column count, i need to refactor these and
add the ability to add contents.

I am requireing the user to set the size explicitly because
I don't want them to pass in a 2D data structure to find
the size from. I feel like that would defeat the perpouse
of the class. Clients will have to set the size manually
then fill it with objects to seperate responsibilities.